requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
stripe
pdfplumber
python-docx
PyPDF2
python-multipart
requests
